preserve command-specific arguments when changing other diff args
Changing the arguments used in a diff buffer did previously cause
command-specific diff arguments to be lost.  For example when the
diff buffer showed the staged changes, then changing the arguments
did also drop the `--cached' argument, resulting in the unstaged
changes to be shown instead.

Now the buffer-local value of `magit-refresh-args' in diff-derived
buffers has the form (REVS CONSTS ARGS FILES).  CONSTS are the
command-specific arguments.  They can only be changes using a diff
command, but not from the refresh popup.