preserve buffer-local diff arguments
Previously `magit-diff-popup' used the value of `magit-diff-arguments'
to initiate the values of its infix arguments.  (Usually the global
value, but due to user errors it could also have been the buffer-local
value).  That made it impossible to change what diff is shown and at the
same time to keep using the previously modified arguments.  (Unless the
modification was explicitly saved, but that caused the value to be used
in all repositories.)

When the `magit-diff-mode' buffer exist for the current buffer, then
`magit-diff-popup' now gets the arguments from that buffer's
`magit-refresh-args', even if that buffer is not current when the popup
is invoked.  Unfortunately that is even the case when the user then goes
on invoking the `magit-show-commit' suffix command, which uses the
`magit-revision-buffer'.  We cannot do anything about that because we
cannot read the users mind.