magit-diff-refresh-popup: allow changing the file filter
Add the `-- FILE...' pseudo argument to the `magit-diff-refresh-popup',
allowing users to change that filter for the current buffer.  Previously
it was not just not possible to do so, but changing another argument
would have removed the file filter.

For buffers that use `magit-diff-section-arguments' add an additional
buffer-local variable `magit-diff-section-file-args' to hold the value
of the file filter.  The refresh suffix command from the popup now
update both variables, and `magit-diff-arguments' includes its value
in the return value when appropriate.

The `magit-insert-{unstaged,staged}-changes' functions respect the new
variable and because the user might otherwise not be aware that a file
filter is in affect in the status buffer, a new section inserter,
`magit-insert-diff-filter-header', is added to the default value of
`magit-status-headers-hook'.

The new function `magit-popup-import-file-args' takes care of combining
the file filters into the pseudo argument form used inside diff popups.