magit-cherry-refresh-buffer: mark arguments as unused
Mark the arguments to `magit-cherry-refresh-buffer' as unused by
prefixing their names with "_".

The `magit-refresh-function' is always called with the buffer-local
value of `magit-refresh-args' as arguments.  This particular refresher
does not use these arguments directly.  Instead it calls the functions
in `magit-cherry-sections-hook' which then use the buffer-local value
of `magit-refresh-args' directly (not as arguments).