make magit-log-select-mode more independent of magit-log-mode
Use a separate buffer for logs used to select a commit (whose name is
controlled by the new option `magit-log-select-buffer-name-format') and
a new dedicated refresher `magit-log-select-refresh-buffer'.

Allow users to set the diff arguments used when selecting a commit from
a log by editing the new dedicated variable `magit-log-select-arguments'.
We shouldn't use the same variable as for regular logs, because it is
likely that users want fancy arguments there, but not here.