delay dropping --graph when it is incompatible with other args
When it is incompatible with other used arguments, then remove `--graph'
from the list of arguments to be used in `magit-log-refresh-buffer'
instead of in `magit-log' (which sets the buffer-local value using
`magit-mode-setup').  This has the advantage that once the incompatible
argument is dropped, `--graph' does not have to be explicitly enabled
again.