magit-log-popup: add --follow argument
Add the `--follow' argument to `magit-log-popup' and when called from
a log buffer, also `magit-log-refresh-popup'.  This argument can only
be used when showing the for a single file, so when the log is not
explicitly limited to a single file, then it is silently dropped.
When `--follow' is used, then `--graph' in turn is silently dropped.