magit-git-command: run git in correct directory
Run git in the directory specified by the DIRECTORY argument.

Do *not* do that by let-binding `default-directory'.  That is
disallowed, because when the `current-buffer' is the process buffer,
then that would blow up.  Instead set that variable in a temporary
buffer.  Luckily we can just reuse the temporary buffer used in this
function for a different purpose, so we don't even have to feel bad
about wasting a buffer.

Fixes #2130.