magit-ediff: run commands from top-level directory
All magit-ediff commands require that the file argument is relative to
the top-level directory.  However, if called from a buffer where
default-directory is not the top-level directory, this results in the
wrong file name being passed to functions like find-file-noselect and
get-file-buffer.

For all magit-ediff commands that have the possibility of using files
from the working tree, set default-directory to the top-level directory.

Fixes #2143