magit-apply-diff: handle diffstat sections
We use the same section type `file' for sections that only show the
name of files and sections whose body is some diff for the file it
represents.

When a `file' section is the child of the `diffstat' section, then fall
back to the respective "diff" `file' section in `magit-apply-diff'.

Fixes #2148.