magit-read-string-ns: new function
Add a new optional argument NO-WHITESPACE to `magit-read-string'.
If it is non-nil then use the new `magit-minibuffer-local-ns-map'
instead of `minibuffer-local-map'.  That map binds "SPC" and "TAB"
to the new command `magit-whitespace-disallowed' which informs the
user that whitespace is disallowed.  If the user manages to insert
whitespace through other means, then raise an error.

Add the new function `magit-read-string-ns' which is a wrapper
around `magit-read-string' which always sets NO-WHITESPACE to t.

Use the new `magit-read-string-ns' in most places where previously
`magit-read-string' was used.

Re #2145.