Double-check that directory is a directory
For some reason, magit blame was failing with exactly one file that I
have, saying that I was not visiting a tracked file.

After some debugging, it looked like magit was setting
`default-directory' to `/path/to/file.py/' instead of `/path/to/'.

Something might be wrong with the way that emacs keeps track of whether
or not a node is a file or a directory on OSX.