magit-insert-head-header: don't hardcode HEAD hash
Despite its name, magit-insert-head-header is used by magit-cherry to
insert information about a commit that is often, but not necessarily,
the repo's HEAD.

Previously, magit-insert-head-header inserted commit information for
HEAD regardless of the BRANCH argument.  As a result, giving
magit-cherry a branch other than the current one as the first argument
resulted in an incorrect header:

  Head:     wrong-commit correct-branch wrong-subject
  Upstream: correct-commit correct-branch subject

To fix this, use the BRANCH argument to retrieve the revision
information if BRANCH is non-nil.