magit-section-goto-successor: avoid falling back to back button
Depending on option `magit-revision-show-xref-buttons' diff buffers may
end with a "back" buttons.  When refreshing a buffer we try to select
the reincarnation of the previously current section and if that fails
fallback to the reincarnation of some "related" section, like the next
sibling or parent.  The final fallback is whatever section came after
the selected section before the refresh and has a reincarnation after
the refresh.  In diff buffers that very often is the "back" button, but
focusing on that section is useless.  Instead we explicitly detect this
case and instead of going to that button go to the beginning of the
buffer.

Re #2164.