magit-get-next-tag: fix handling of nil rev
Although 'git describe' is the same as 'git describe HEAD',
'git describe --contains' does not output anything and is not the same
as 'git describe --contains HEAD' (at least with git 2.5.0).

magit-insert-tags-header calls magit-get-next-tag with nil as the REV
argument, so the next tag was not being inserted even when it existed.