magit-blame: fix calls from indirect buffers
Make sure that default-directory stays set at the top-level directory.
When magit-blame is called from an indirect buffer, the find-file call
sets default-directory to the file's directory, so the downstream git
call (now from the file's directory) fails when given a path relative to
the top-level directory.

This isn't an issue when magit-blame is called from a buffer visiting a
file.  The call to find-file does not override the default-directory
let-bound by magit-with-toplevel because the file is already being
visited by the current buffer.

Fix this by let-binding the find-file call separately.

While this makes it possible call magit-blame from an indirect buffer,
point is not placed appropriately in the base buffer.  This probably
isn't worth taking the time to fix given that calling magit-blame from
indirect buffer isn't very common (since the above error has yet to be
reported).