magit-diff-wash-diffstat: use --numstat to get full file names
This is a kludge which replaces the older kludge removed in
c64bb994d905d0daf3162abce4c425e6c097b8d2.  It fixes #2199.

It would be better if we did instead use *only* `--numstat' and
then produced output similar to what `--stat' does, but without
actually relying on output from that.