with-editor-return: use server-done instead of server-edit
This fixes #2197.  We don't actually want to show the buffer belonging
to another client, so we call `server-done' instead of `server-edit'.
The latter calls the former and then also tries to select a buffer,
using `server-switch-buffer' which favors buffers belonging to another
client.  This does matter when Emacs was started in daemon mode and the
user connects to the instance using the emacsclient and then uses Magit
to invoke.  When s/he then completes the commit, then their always is
another client, the one used to connect to the instance in the first
place.

Later in `with-editor-return' we restore the window configuration from
before `with-editor' was used, so one would think that needlessly
showing another buffer before restoring the window configuration would
not matter.  However the previous window configuration has to be
explicitly stored by the caller of `with-editor', so we cannot rely on
a configuration being stored and have to fix the above issue anyway.

In the case of committing using Magit this should not have made a
difference because that does save the previous window configuration, but
unfortunately it does it to late, when the configuration has already
been modified somewhat.  I am not sure if that can be fixed, but in any
case that's a different issue.