magit-rev-format: always treat REV as a commit, add ARGS
This function is intended to print information about a commit, so cast
it as a commit by appending "^{commit}" to REV.  This is useful when REV
actually is a tag, in that case `git show' always prints tag related
information before the information requested using FORMAT, and so this
function used to return incorrect information in that case.

If the return value would be an empty string, then instead return nil.

Also add the new argument ARGS.  These arguments are passed to `git
show' and should be used for arguments such as `--date=' which are
respected by certain placeholders in FORMAT.