magit-blame-popup: use read-string for -C and -M
These arguments expect a number as value, but the value can also
be omitted in which case git uses some default value.  By using
`magit-popup-read-number' we forced the user to input a number.

Now we use just `read-string', allowing the user to input a number,
nothing, or something invalid (which unfortunately won't be noticed
until git tries to use it).

Fixes #2208.