magit-diff-wash-diff: fix diff header regexp
The diff header looks like on of these:
  diff --git a b
  diff --cc a
  diff --combined a

Change the regexp so that both "a" and "a b" fall into the second
submatch.  For `--git' that means that the variables `orig' and `file'
are set to wrong values, but that does not matter, since in that case
they are later set again from other headers.

Fixes #2211.