jump to next block of changed lines after staging region
After using the region to stage or discard parts of a hunk, jump to
the beginning of the next block of changed lines.  After staging the
last block, instead jump to the previous block.  This is implemented
using regexp searches, not by counting lines.

Previously we just jumped to the beginning of "the" hunk (staging a
region may add or remove a hunk), but that can be very far away from
where point previously was.

It might be better to jump to "the next line which still exists",
but it would be rather complicated to figure out what line that is.