magit-toplevel: refactor and add tests
The idea is to only follow symlinks if absolutely necessary.  The
previous implementation was broken in that it usually failed when
a symlink to some sub-directory of a repository was involved (and
when that succeeded, then only by accident).

This issue was reported in #2135.  A previous attempt (#2209)
failed to address this issue, as it caused other cases to fail.