magit-insert-am-sequence: check that commit exists
The previous version failed if a commit in the patch series didn't
already exist as a commit object (magit-format-rev-summary returned nil,
which magit-sequence-insert-commit tried to insert).

Rewrite magit-insert-am-sequence to check whether the commit from the
"From" line of the patch exists.  Insert a section for the commit if it
does and a section for the patch file otherwise.