call magit-mode-get-buffer with nil as FORMAT
`magit-mode-get-buffer' can derive the name of the variable whose value
is the buffer-name format string to be used based on the MODE argument.
Some callers already make use of this and pass nil as FORMAT, but many
others did not do so until now.

Deriving the name of the variable from the name of the mode is fairly
cheap so it is okay to do so, especially because nor specifying the
format variable explicitly makes many callers cleaner.  However we
cannot completely remove the FORMAT argument for backward (and possibly
forward) compatibility reasons.