magit-toplevel: rename argument FILE to DIRECTORY
This can actually be a regular file, but that should not be advertised.

To avoid setting a bad example, make sure `magit-file-relative-name'
calls `magit-toplevel' with a directory as DIRECTORY, even though that
is not technically necessary, and it does so be doing the exact same
thing `magit-toplevel' would do anyway.