magit-mode-setup: don't create buffer outside repository
Previously `magit-mode-setup' first created the buffer and only then
did it check whether it was actually called from within a repository.

Also when `find-file-visit-truename' is non-nil and `magit-mode-setup'
is called outside any repository, then `file-truename' was called on
`nil', resulting in a type error instead of the much more appropriate
"Not inside a Git repository".