Revert "support stashing when there are conflicts"
This reverts commit 41965a9f06b4d1737fe81a728da3b0c7c6cd4b3c.

We had to be able to create stashes while there are conflicts because
the `magit-backup-mode' used stashes for the backups.  But since that
mode was replaced by the `magit-wip-*' modes, we no longer need this.
Git doesn't allow creating stashes while there are conflicts and now
doesn't do so anymore either.