address issues concerning `git read-tree'
The macro `magit-with-temp-index' used `git read-tree' to create a
temporary index file.  Only reading into the index without performing a
merge is expensive, so it is preferable to use `-m' to perform a merge.
But then the command would refuse to run if there are any conflicts.

The needs of the various users of this macro are different and so they
need to be able to specify whether `-m' is should be used or not, and
whether any other arguments should be used.

Add a new argument ARG to `magit-with-temp-index' as well as to
`magit-commit-worktree'.  In both cases this can be a single argument or
a list of arguments (but we don't need that currently).  Then adjust the
users.

When creating a stash `magit-stash-create' uses `magit-with-temp-index'
twice.  In one case the TREE is nil, so ARG is irrelevant and we pass
nil for that too.  In the other case we pass `-m' to avoid the slowdown
and also because we want to error out if there is a conflict.

When creating a wip commit `magit-wip-commit-index' now passes `--reset'
to `magit-with-temp-index' to avoid the slowdown and to add the conflict
to the index, pretending it was resolved.  If there is a conflict then
it is staged including the conflict markers.  This actually is what we
want, we want to save the current state of the *file* (not the index),
including possibly partially resolved conflicts.

In `magit-sequence-insert-sequence' a dummy commit is created using
`magit-commit-worktree'.  It is used to figure out whether the changes
introduced by the stopped at commit still exists in the worktree.  Here
we use `--reset' to avoid the slowdown and in case there are conflicts
any errors.

Re #2049.