get the rebase arguments inside interactive again
Previously we tried moving that task which is common to all rebase
variants to `magit-rebase-interactive-1', but since that isn't called
inside the commands' `interactive' form, it does not work reliably.
Most importantly when such a command is invoked directly instead of
from within the popup, then the arguments would not be used.