magit-section-goto-successor: fix hunk header case
In af5877ed0d68f16770184554674a3139b2d30df3 a special case was added
to jump to the next block of changed lines after staging a region.

But that also removed some code which is necessary when staging the
first hunk with point on the header.  Restore that code, but place it
inside the code added for the special case mentioned above.

Fixes #2267.