magit-rebase-popup: change two bindings
Bind "i" instead of "e" to `magit-rebase-[i]nteractive'.
Bind "e" instead of "s" to `magit-rebase-[e]dit-commit'.

Such changes are often not well received, so here we go:

Back in the days before I even used Magit, "r" was bound to a primitive
rebase command and "e" was bound to a primitive interactive rebase
command, probably "i" was already used for something else.  Eventually I
improved these rebase variants and added additional variants, all of
which are usually being invoked from the new rebase popup.  Both "r" and
"e" were bound to the popup command to make the transition easier.

So it was possible to press "rr" to start a non-interactive rebase and
"ee" for an interactive rebase (of course "er" and "re" would have
worked too).  But then some new, unrelated commands were added and we
needed to find keys that could be bound to those.  That's how the rebase
popup lost the privilege of being the only command which has two
top-level bindings, and only "r" remained.

Inside the popup we kept "e" bound to the interactive rebase variant,
because, as you are probably aware of if you read this, changing key
bindings lets muscle memory go to waste and some people find that very
disturbing.

But because in this case maintaining the old bindings leads to many
inconsistencies, I have decided to fix it now - the complains will pass.

"interactive" begins with "i" not "e".  The rebase popup also features
an "--interactive" switch in addition to the "interactive" command (this
switch allows turning the non-interactive actions into interactive
actions, which can be useful in some cases).  That switch's binding is
"-i" not "-e".  The popup also features a "edit only this particular
commit" variant `magit-rebase-[e]dit-commit' which originally "s" was
bound to because "e" was taken (cannot remember why "s", probably
because it is used to change a [s]ingle commit).  That command was
invoked using "rs" and was a shorthand for "ree".  The "s" and "e" serve
the same purpose here so they should be the same key, and after this
change they are: "re" and "rie".