magit-commit-diff: respect diff arguments
If the diff arguments aren't passed, the return value of
magit-diff-arguments is unintentionally affected in future calls.  For
example, even if the magit-diff-arguments return value included --stat
before the commit, a revision buffer created after a magit-commit-diff
call will not have a diffstat section (assuming the diff buffer is still
around and the arguments haven't been explicitly changed since the
commit).

Drop file restrictions so that the commit diff always shows the full set
of changes.  The act of committing probably shouldn't reset file
restrictions that would otherwise be carried over between non-commit
diff calls, but this was already the situation before this change, and
it seems better than limiting the commit diff to a set a files from a
previous diff, which is unlikely to be what the user intended.