server-switch-buffer(with-editor): save window configuration
Save the window configuration in the `with-editor' advice to
`server-switch-buffer', but only if `with-editor-mode' is enabled.

Previously saving the window configuration was left to callers, but
by the time they got a change to do so, it was already to late, as
the client buffer was already being displayed.  Also we only did
this for committing but not rebasing.

Fixes #2273.