Use the "real" git of Git for Windows 2+
Git for Windows 2.x reorganized the directory structure so both cmd/git
and bin/git are just wrappers that call the real thing.  Try to find and
use the "real" one instead (it's faster) by executing `which' via the
wrapper.  We also need to set PATH so that git can find all its
utitilities and dlls, this can be retrieved from the wrapper as well.

close #2259