magit-pull-{,current}: use magit-run-git-with-editor
While the `magit-pull-popup' does not feature the `--edit' switch, a
user might have added that herself or set GIT_MERGE_AUTOEDIT to yes,
so we have to make sure git knows how to call home.  To do that use
`magit-run-git-with-editor' instead of `magit-run-git-async' as we
usually do in such cases.

Fixes #2285.