magit-log-refresh-buffer: disable graph speed hack for some args
When calling `git rev-list' to determine the number of commits, so that
`git log --graph' can later be called with a limited range to avoid
needlessly calculating the graph for the complete history, use all the
specified arguments.  Otherwise we might end up limiting the history too
much.

When certain arguments are used, that would actually double the time it
takes to determine the commits to be shown (because we would be doing it
twice), which by far outweighs what we gain from avoiding to needlessly
calculate the complete graph.

Add new variable `magit-log-disable-graph-hack-args' which lists these
arguments and consult it in `magit-log-refresh-buffer' when determining
whether the hack should be used or not.

Fixes #2287.