use same washer function for oneline and verbose log variants
Remove the washer functions `magit-log-wash-verbose' and
`magit-log-wash-line' (which was used for all other variants beside
"verbose") and replace them with `magit-log-wash-rev' (which is now
used for all variants).

That's not very nice, this function is turning into a monstrosity and
it's only going to get worse over the next few commits.  Eventually
this has to be split up, but right now combining the two variants is
the right thing to do.

The verbose log now shows the same oneline summary as the "oneline"
variant instead of something resembling `--format=medium'.  This
will come back in a different form in a few commits (in the form of
the `++header' argument).  The term "verbose" now only refers to the
use of an argument like `--patch' or `--stat' (i.e. we should stop
talking about "verbose" logs).