don't use a separate variable for the `-n' log argument
Add the `-n' argument to the `magit-log-popup' and add "-n256" to the
default value of `magit-log-arguments', and in turn remove the variable
`magit-log-cutoff-length'.

This will lead to invalid reports about a performance regression in the
log buffer.  But what actually happened is that the user customized the
global value of `magit-log-arguments' and that value did not get updated
when the default values was changed.  Instruct users to add "-n256" to
the customized value.

This change is motivated by the drastic difference in showing a log with
and without diffs.  It should therefore be possible to quickly increase
and decrease the number of shown commits, without having to customize a
variable manually (i.e. not from the popup or using a dedicated
command).

Also add new the commands `magit-log-toggle-commit-limit',
`magit-log-double-commit-limit', and `magit-log-half-commit-limit' to
make this even easier, and replacing `magit-log-show-more-commits'.

For reflogs we no longer impose a limit because they never come with
graphs and washing them is quick.