magit-log-{popup,buffer-file}: add `-L' argument
Support is pretty basic and will be improved eventually.  It won't be
improved incrementally, i.e. I won't add related arguments as they are
being requested, but only when I feel that the time to do so has come,
i.e. once I have the time to tackle this exciting area of Git properly.