use regular `commit' section type in stash buffers
This is necessary to allow creating patches from the individual
commits that form the stash.  Also there's not really any benefit
in using different section types for the sections in stash buffers.

Also change the headings of these commit sections - now that the
header-line shows the stash number, that information no longer has
to repeated for each commit section.