magit-popup-mode-setup: add kludge to force proper window start
In v24.4 `fit-window-to-buffer' seems to "fit to buffer, except for the
whitespace at the end of it".  In v25.0.50 it does even scroll the
window down so that the beginning of the buffer is no longer visible in
the window.

By counting the lines in the buffer and passing that as the MIN-HEIGHT
to `fit-window-to-buffer', we avoid these issues.