apply: avoid more unnecessary refreshes and wip saves
Remove the variable `magit-apply-inhibit-wip'.  Instead of consulting
that variable `magit-apply-patch' now consults `inhibit-magit-refresh',
and if that is set it neither refreshes nor saves to the wip refs.

The end result is the same, except that `magit-reverse-files' now not
only avoids needlessly saving to the wip refs, but also avoids repeated
refreshes.