magit-mode-{setup,display-buffer}: re-calibrate responsibilities
Determine, and if necessary create, the buffer inside `magit-mode-setup'
instead of inside `magit-display-buffer'.

For both functions the first argument previously had to a buffer or a
buffer-name format string.  Now `magit-mode-setup' expects a format
string and `magit-mode-display-buffer' expects a buffer.  Additionally
`magit-mode-display-buffer's MODE and TOPDIR arguments are removed
(they were only necessary in order to create the buffer), it no longer
performs any sanity checks, and it no longer returns the buffer (the
callers already know the buffer).