magit-mode-setup: derive the buffer-name format from the mode
Remove the FORMAT argument.  Instead derive the variable whose value
should be used as the format string used to create the buffer-name from
the MODE argument.  The code which does this is already in place in
`magit-mode-get-buffer-create', so `magit-mode-setup' just has to call
that with `nil' as FORMAT.

Also switch the positions of `magit-mode-setup's MODE and SWITCH-FUNC
arguments.