remove DIRECTORY argument of magit-mode-get-buffer et al.
Remove `magit-mode-get-buffer's PWD argument,
remove `magit-mode-get-buffer-create's DIRECTORY argument, and
remove `magit-process-buffer's PWD argument.

The only real purpose these arguments served was that this allowed not
raising an error when `magit-process-buffer' requested a buffer outside
any repository.

The place where we assert that we are inside a repository moves from
`magit-mode-setup' to `magit-mode-get-buffer' and we now explicitly
check whether the mode is `magit-process-mode' in which case we don't
raise an error.  This allows removing the confusing "directory"
arguments.