derive name of refresh function from mode name
Remove the buffer-local variable `magit-refresh-function' and
`magit-mode-setup's REFRESH-FUNC argument, which was used to set the
former.

Instead `magit-refresh-buffer' determines the name of the refresh
function based on the name of the mode.  If that function is defined,
then it is used to refresh the buffer, otherwise the buffer is not
refreshed.

This is possible because the names of all refreshers consistently follow
this naming schema.  Not passing the value around has the advantage that
the code gets simpler and it also helps enforcing this convention.

`magit-process-mode' is currently the only mode without a corresponding
refresher function.