magit-process-buffer: don't use magit-mode-get-buffer-create
We already did not use `magit-mode-setup' in order to avoid unnecessary
work every time a process is run for side-effects and because that would
run `magit-mode-hook' which by default results in git being run, which
could have lead to an error (see below).

Now `magit-process-buffer' also finds or else creates the buffer itself,
instead of using `magit-mode-get-buffer-create' for that purpose.

If `default-directory' isn't inside a repository, then it walks up the
directory until it finds "the" root directory (usually "/", but when
using Tramp or MS Windows, then something else), and finds or creates
the buffer for that directory.  Even if the root directory isn't also
the toplevel of a repository (which it seldom is) it is never-the-less
used in the buffer name, e.g. *magit-process: /*, and as the
`default-directory' of that buffer.

Previously when this function was called for the first time with a
`default-directory' that isn't located inside a repository, then a
buffer *magit-process: -* was create and the `default-directory' which
was current when this function was called was used as the buffer-local
value.  That value was never changed, which was a problem if that
directory was ever deleted (e.g. the `magit-mode-hook' issue mentioned
above).