uniquify names of magit buffers by default
Add new option `magit-uniquify-buffer-names' with a default value of t.

Due to how `uniquify.el' is implemented, buffer name cannot be made
unique before they have been created, so part of this has to be
implemented in `magit-generate-new-buffer'.

Modes are added to `uniquify-list-buffers-directory-modes' on demand
instead of upfront, because this way derived modes implemented in
third-party extensions are automatically taken care of.