run magit-mode-setup-hook after displaying the buffer
In [1: 9eb318b] the order of of events in `magit-mode-setup' was
changed from

  1. display buffer
  2. run setup hook
  3. turn on mode
  4. refresh buffer

to

  2. run setup hook
  3. turn on mode
  1. display buffer
  4. refresh buffer

because `magit-display-buffer's behavior now depends on the major-mode.
I considered removing `magit-mode-setup-hook' completely because given
the above order of events it is not necessary, one could just use
`magit-mode-hook' or a more specific `magit-{*}-mode-hook'.

But it turns out we actually need a hook that runs after displaying the
buffer (see next commit).  So change the order to

  3. turn on mode
  1. display buffer
  2. run setup hook
  4. refresh buffer

By the way, it would be preferable if we could use this order

  3. turn on mode
  1. display buffer
  4. refresh buffer
  2. run setup hook

but that is currently not possible because some section inserters only
work properly if the buffer the insert into, is being displayed in a
window.  Also note that `magit-display-buffer' has to run outside any
`with-current-buffer' or `safe-current-buffer' forms.

Also define `magit-mode-setup-hook' as an option instead of as an
undocumented variable.

1: 9eb318b02a5d2df127c213ab11a3ff199a7e9301
   concentrate display-buffer logic in one place