assert that remote git can be found and is recent enough
In `magit-status-internal' call the new `magit-tramp-assert' to make
sure that when it shows the status of a remote repository using Tramp
git can be found on the host and that it is recent enough.

For local repositories `magit-tramp-assert' bails quickly and is not
very expensive.  Once `magit-tramp-assert' has found a good git for a
given host, then it caches that information.

The assert is only performed in `magit-status-internal' because it
(`magit-status' actually) it the main entry point.