magit-diff-wash-diffstat: prevent invalid hash
As of 36fb789 (populate revision buffers using section inserter
functions, 2015-08-27), making a log with --stat and without --graph
results in a log like

    19a19a6 magit-maybe-save-[...]
    1 file changed, 2 insertions(+), 1 deletion(-)
    lisp/magit-mode.el | 3 ++-

 -> 26e688 improve [...]
    3 files changed, 30 insertions(+), 19 deletions(-)
    Documentation/magit.org  |  4 ++--
    Documentation/magit.texi |  4 ++--
    lisp/magit-mode.el       | 41 ++++++++++++++++++++++++++---------------

where the first characters of all hashes following the first entry are
deleted.

Instead of deleting a character after inserting a diffstat, move forward
if looking at a blank line (which is true in revision and diff buffers).
Otherwise, insert a newline (which is true in non-graph log buffers with
--stat).