magit-commit-show-diff: new option
Add new option `magit-commit-show-diff' to control whether a diff for
the relevant changes is shown while.  A value of `t' has the same effect
as `commit' and `log-select' being members of `magit-diff-auto-show' had.

Also add a third value, `verbose', to the option `magit-commit-ask-stage',
as a subsitute for the old `magit-commit-show-diff' member `stage-all'.