magit-diff-auto-show: drop log-oneline variant
Previously `log-oneline' was the only remaining member of
`magit-diff-auto-show'.  It is being replaced with `log-follow'.

`magit-log-maybe-show-commit' now only ever updates the revision
buffer being displayed in another window of the same frame.  It
does not automatically displays the buffer in another window if
it is not already being displayed.  The user now has to create
the buffer and window once, by pressing `SPC'.