separate updating of revision and blob buffers
Add new hook functions `magit-log-maybe-update-revision-buffer'
and `magit-log-maybe-update-blob-buffer', and add the former to the
hook `magit-section-movement-hook'.  They replace the hook function
`magit-log-maybe-show-commit' which previously was a member of that
hook.

Remove the variable `magit-update-other-window-timer' variable, the two
functions now use separate variables `magit--update-revision-buffer' and
`magit--update-blob-buffer' to prevent multiple timers from running in
parallel.  There may be two timers now, one for the revision buffer and
one for the blob buffer.

Remove the option `magit-diff-auto-show' is removed.  To disable the
automatic updates of the revision buffer, users now have to remove a
hook function from `magit-section-movement-hook', and to enable updates
of the blob buffer, they have to add a function to that hook variable.

Define `magit-section-movement-hook' as an option.

Due to the removal of `magit-update-other-window-timer' the function
`magit-blame-maybe-update-revision-buffer' has to be adjusted too.